requirements.txt
python-dotenv
openai
PySimpleGUI
sounddevice
soundfile
